Switching to the ESP32-P4-ETH board

+++ src/led/LedController.h
@@ -11,15 +11,16 @@ namespace led {
  * @brief GPIO LED controller implementation.
  *
  * Controls an LED connected to a GPIO pin (default GPIO 26 / D26).
- * Active high by default.
+ * Active high by default. Passing -1 disables the LED (no-op) for
+ * boards without a controllable user LED (e.g. ESP32-P4-ETH).
  */
 class LedController : public ILedController {
 public:
     /**
-     * @param gpioPin   GPIO pin number (default 26).
+     * @param gpioPin   GPIO pin number (default 26, or -1 to disable).
      * @param activeHigh true = GPIO high turns LED on (default).
      */
-    explicit LedController(uint8_t gpioPin = 26, bool activeHigh = true);
+    explicit LedController(int gpioPin = 26, bool activeHigh = true);
     ~LedController() override;
 
     void turnOn() override;
@@ -28,7 +29,7 @@ public:
     void toggle() override;
 
 private:
-    uint8_t gpioPin_;
+    int gpioPin_;
     bool activeHigh_;
     bool on_ = false;
 };
